Import live.2013.04.08.tar.gz
2013.03.31
- Added support for decoding '%<hex><hex>' sequences if they exist in RTSP URL stream names.  (This allows media server file
  names to contain spaces, or non-ASCII UTF-8 characters, for example.)  (Thanks to Warren Young for proposing this.)
- Our "ProxyServerMediaSession" code currently does not support the proxying of JPEG video or AMR audio streams
  (because the data output by the "RTPSource" object is not in a form that can be fed directly into the corresponding
  "RTPSink" object).  This may be fixed sometime in the future, but, in the meantime, we output an error message
  (when in 'verbose' output mode) if we try to proxy such streams.

2013.04.01
- Our "ProxyServerMediaSession" code can now proxy JPEG video RTP streams.  (It does so by copying the raw JPEG/RTP payloads from
  a "SimpleRTPSource" to a "SimpleRTPSink", without interpreting the JPEG-specific header fields at all.  We also had to add
  a hack for 'copying' the RTP 'M' bit as well.)
- Some minor changes to eliminate some compiler warnings.

2013.04.04
- Because at least one IP camera out there seems to crash whenever it receives a "GET_PARAMETER" command, even though it
  reported - in response to our earlier "OPTIONS" command - that it supported "GET_PARAMETER", we updated the
  "ProxyServerMediaSession" code to send "GET_PARAMETER" as a 'liveness' command only if the server earlier reported
  (in the "SETUP" response's "Session:" header) a non-zero "timeout" parameter.  (Thanks to Roman Gaufman for demonstrating
  such a buggy server.)

2013.04.05
- Updated the "RTSPClient" implementation of "GET_PARAMETER" to handle response stringss that don't begin with the parameter name.
  In this case, we return the entire result string.

2013.04.06
- Updated the change to the proxying code that we made in version 2013.04.04.  By default, we now *never* send "GET_PARAMETER"
  as our 'liveness indicator' command, because some IP cameras seem to crash whenever they receive "GET_PARAMETER".
  (The old code that sent "GET_PARAMETER" is still there, but #ifdef'd out, in case anyone wants to send "GET_PARAMETER".)
- Increased "OutPacketBuffer::maxSize" in the "live555ProxyServer" application to 100,000 bytes, to accommodate some camera servers
  that send ridiculously large frames.  (Note, however, that if the back-end network (from the proxy to clients) has any significant
  packet loss, then these ridiculously large frames might not end up getting delivered to clients.  Instead, you should fix your
  back-end server to not send frames this large.)
- Updated the "RTPReceptionStats" code (in "liveMedia/RTPSource.cpp") to eliminate the possible use of some uninitialized variables
  (if a RTCP "RR" report is sent before we have received any RTP packets).  (Thanks to Michel Promonet for noting this.)

2013.04.08
- Some systems needed to #include <ctype.h> in "RTSPCommon.cpp" to compile "isdigit()", used by the code that we added in
  version 2013.03.31.  (Thanks to Michel Promonet for noting this.)

+++ liveMedia/OnDemandServerMediaSubsession.cpp
@@ -228,7 +228,7 @@ void OnDemandServerMediaSubsession::seekStream(unsigned /*clientSessionId*/,
   if (streamState != NULL && streamState->mediaSource() != NULL) {
     seekStreamSource(streamState->mediaSource(), seekNPT, streamDuration, numBytes);
 
-    streamState->startNPT() = seekNPT;
+    streamState->startNPT() = (float)seekNPT;
     RTPSink* rtpSink = streamState->rtpSink(); // alias
     if (rtpSink != NULL) rtpSink->resetPresentationTimes();
   }
@@ -277,7 +277,7 @@ float OnDemandServerMediaSubsession::getCurrentNPT(void* streamToken) {
 
     return streamState->startNPT()
       + (rtpSink->mostRecentPresentationTime().tv_sec - rtpSink->initialPresentationTime().tv_sec)
-      + (rtpSink->mostRecentPresentationTime().tv_sec - rtpSink->initialPresentationTime().tv_sec)/1000000.0;
+      + (rtpSink->mostRecentPresentationTime().tv_sec - rtpSink->initialPresentationTime().tv_sec)/1000000.0f;
   } while (0);
 
   return 0.0;
